Refactor: subdirectory for editor widgets
