+++ lotti/lib/pages/create/create_task_page.dart
@@ -6,7 +6,7 @@ import 'package:lotti/classes/task.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/widgets/journal/editor_tools.dart';
+import 'package:lotti/widgets/journal/editor/editor_tools.dart';
 import 'package:lotti/widgets/tasks/task_form.dart';
 
 class CreateTaskPage extends StatefulWidget {
