+++ lotti/lib/widgets/journal/entry_details_widget.dart
@@ -15,8 +15,8 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/services/editor_state_service.dart';
 import 'package:lotti/theme.dart';
 import 'package:lotti/widgets/audio/audio_player.dart';
-import 'package:lotti/widgets/journal/editor_tools.dart';
-import 'package:lotti/widgets/journal/editor_widget.dart';
+import 'package:lotti/widgets/journal/editor/editor_tools.dart';
+import 'package:lotti/widgets/journal/editor/editor_widget.dart';
 import 'package:lotti/widgets/journal/entry_details/entry_detail_footer.dart';
 import 'package:lotti/widgets/journal/entry_details/entry_detail_header.dart';
 import 'package:lotti/widgets/journal/entry_image_widget.dart';
