+++ lotti/lib/widgets/tasks/task_form.dart
@@ -9,7 +9,7 @@ import 'package:lotti/classes/task.dart';
 import 'package:lotti/theme.dart';
 import 'package:lotti/utils/task_utils.dart';
 import 'package:lotti/widgets/form_builder/cupertino_datepicker.dart';
-import 'package:lotti/widgets/journal/editor_widget.dart';
+import 'package:lotti/widgets/journal/editor/editor_widget.dart';
 import 'package:lotti/widgets/tasks/linked_duration.dart';
 
 class TaskForm extends StatefulWidget {
